requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptocode==0.1
cryptography==38.0.1
Django==4.1.1
django-environ==0.10.0
djangorestframework==3.14.0
docopt==0.6.2
elastic-transport==8.4.0
elasticsearch==8.4.2
elasticsearch7==7.17.6
filetype==1.1.0
geographiclib==1.52
geopy==2.2.0
idna==3.4
localStoragePy==0.2.3
minio==7.1.12
pipreqs==0.4.11
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.15.0
pyindia-zipcode==0.0.2
PyJWT==2.5.0
pymongo==4.2.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
yarg==0.1.9
zipcodes==1.2.0
django-redis==5.3.0
redis==4.6.0
pika==1.3.2
